Changed task parameter handling, TaskAccumulator math, and the test TaskFoo now is an actual PD controller.
Details:

Refactored the task parameter idea into a semi-abstract base class
with concrete subclasses for each supported type. Each parameter
stores just a pointer to the actual variable, which can live wherever
the task implementers choose. For instance, right in their task
subclass. This makes it much more lightweight in terms of usage. Maybe
it can be made a bit more versatile yet using templates, but the need
to declare setters and getters for each supported type in the
parameter base class makes this a bit moot.

Looked over attic code and found that Luis uses a Jacobian projected
via the dynamically consistent nullspace when computing desired task
accelerations. Not quite sure whether I got it right though, so the
compiler warning statement remains until we've tested it.

The testTask now incorporates a "real" controller, which can now be
used to test the TaskAccumulator. Should seriously start using gtest
for all this...